rename: MarketSupport 클래스 패키지 이동 및 관련 클래스 변경 - #78
